requirements.txt
numpy>=1.21.0
pandas>=1.3.0
networkx>=2.6.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
pyyaml>=5.4.0
tqdm>=4.62.0
pytest>=6.2.0
pytest-cov>=2.12.0
